Clear the last WebGL errors: attribute type and unpack row length
particlesystem.cpp declared its texture coordinates as GL_INT, which is not a
valid vertex-attribute type in WebGL/GLES2 — the driver rejected every
glVertexAttribPointer for the particle buffer. They are texture pixel
coordinates well inside float's exact integer range, so Vec2f with GL_FLOAT is
both valid everywhere and the same 8 bytes, leaving the vertex layout unchanged.

texture.cpp uploads sub-rectangles by setting GL_UNPACK_ROW_LENGTH to the SDL
surface pitch. WebGL 1.0 has no such parameter: it raises INVALID_ENUM and
ignores it, so a surface whose pitch exceeded the texture width would have
uploaded skewed. SDL's 32-bit surfaces happen to be tightly packed, so it never
bit, but the shim now tracks the value and repacks the rows itself on any upload
where it actually differs, rather than leaving that to luck.

A run now reports zero GL errors, and menu, level select and gameplay all render
correctly in headless Chromium.

+++ Blocks5/src/particlesystem.h
@@ -42,7 +42,10 @@ private:
 	struct Vertex
 	{
 		Vec2f position;
-		Vec2i uv;
+		// Float, not int: GL_INT is not a valid vertex-attribute type in WebGL /
+		// GLES2, and these are texture pixel coordinates well inside float's exact
+		// integer range. Same 8 bytes, so the vertex layout is unchanged.
+		Vec2f uv;
 		Vec4f color;
 	};
 #endif
